Cleanup and Revision of Main File
This adds/creates a new version of the compile file which actually
builds the textbook.

This file was based originally on the "splits" version, that is, version
2. It includes significantly more commenting and explanation of the
Latex Code-chunks that are in the main file.

Additionally there are some tweaks made to the wrap-up, copyright info
and other minor things in the pre-review pages.

Add's the "short" command version to all of the margin comment options
to allow some cleaner pages.